refactor: extract Cli struct and routing to src/cli.rs
Move Cli struct definition and command routing logic from main.rs
to a new cli.rs module with a public run() function. main.rs is
now a thin entry point (~45 lines) containing only module
declarations and the tokio::main entry point.

Closes #59
